Add feature specs for closed events dashboard
* Additionally, refactor helper methods from current events dashboard
spec into a module in spec/support for reuse
