+++ spec/rails_helper.rb
@@ -18,6 +18,7 @@ RSpec.configure do |config|
   config.include FactoryGirl::Syntax::Methods
   config.include Monban::Test::ControllerHelpers, type: :controller
   config.include UserReviewHelpers, type: :feature
+  config.include EventSummaryHelpers, type: :feature
 
   config.after do
     Monban.test_reset!
